lib::wavelet; restructure, module munging
- move Wavelet and Bin to top level
- spec only contains specs & Shape
- stencil was replaced with restrict

Super messy changes.  Got interrupted and lost track of whatever I was doing.
In the middle of overhauling restriction.  Someday, all history is ancient
history.  Move forward and it will be ancient faster.
